Also use BigWord2VecMatrix during transform
This avoids having to pull all word vectors to the driver after the
training and having to broadcast them to all executors during
transformation. Now the Glint cluster is started on Spark also during
transformation and only the word index has to be broadcasted. The final
remaining scalability limitation is the 8GB broadcast size limit for
the word index.

This commit also adapts the integration tests to use a single Spark
session because stopping the session seems to not properly clean up,
resulting in OOMs as described in [SPARK-13198].